添加预览静态页面方法

+++ core/JavaSource/com/fantasy/swp/service/PageBeanService.java
@@ -15,6 +15,7 @@ import org.springframework.stereotype.Component;
 import org.springframework.beans.factory.annotation.Autowired;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 
 
 @Component
@@ -32,6 +33,10 @@ public class PageBeanService {
     @Autowired
     private FileManagerFactory fileManagerFactory;
 
+    /**
+     * 页面列表
+     * @return
+     */
     public List<IPage> listPage(){
         List<Page> pages = pageService.find(new ArrayList<PropertyFilter>());
         List<IPage> ipages = new ArrayList<IPage>();
@@ -43,14 +48,41 @@ public class PageBeanService {
         return ipages;
     }
 
+    /**
+     * 保存
+     * @param website 站点
+     * @param url 页面存储相对路径
+     * @param templatePath 模板
+     * @param name 页面名称
+     * @return
+     */
     public IPage savePage(Website website,String url, String templatePath, String name){
         return this.savePage(website, url, templatePath, name,null,0);
     }
 
+    /**
+     * 保存
+     * @param website 站点
+     * @param url 页面存储相对路径
+     * @param templatePath 模板
+     * @param name 页面名称
+     * @param datas 构成页面数据
+     * @return
+     */
     public IPage savePage(Website website,String url, String templatePath, String name,List<Data> datas){
         return this.savePage(website, url, templatePath, name,datas,0);
     }
 
+    /**
+     * 保存
+     * @param website 站点
+     * @param url 页面存储相对路径
+     * @param templatePath 模板
+     * @param name 页面名称
+     * @param datas 构成页面数据
+     * @param pageSize 分页时pageSize
+     * @return
+     */
     public IPage savePage(Website website,String url, String templatePath, String name,List<Data> datas,int pageSize){
         if(datas!=null && datas.size()>0){
             for(Data data : datas){
@@ -73,6 +105,12 @@ public class PageBeanService {
         return pageBean;
     }
 
+    /**
+     * 获得某个页面
+     * @param url 页面相对路径
+     * @param websiteId 站点id
+     * @return
+     */
     public IPage get(String url, Long websiteId) {
         PageBean pageBean = new PageBean();
         Page page = this.pageService.findUniqueByPath(url,websiteId);
@@ -80,10 +118,20 @@ public class PageBeanService {
         return pageBean;
     }
 
+    /**
+     * 删除
+     * @param url
+     * @param websiteId
+     */
     public void remove(String url, Long websiteId) {
         this.pageService.deleteByPath(url,websiteId);
     }
 
+    /**
+     * 创建页面
+     * @param page
+     * @return
+     */
     public List<IPageItem> execute(Page page) {
         Page newpage = this.generate.create(page);
         List<IPageItem> pageItems = new ArrayList<IPageItem>();
@@ -95,8 +143,14 @@ public class PageBeanService {
         return pageItems;
     }
 
+    /**
+     * 获取页面详情
+     * @param path
+     * @param pageId
+     * @return
+     */
     public IPageItem getPageItem(String path, Long pageId) {
-        PageItem pageItem = pageItemService.findUniqueByPath(path,pageId);
+        PageItem pageItem = pageItemService.findUniqueByPath(path, pageId);
         if(pageItem==null){
             return null;
         }
@@ -106,16 +160,29 @@ public class PageBeanService {
     }
 
     /**
-     * 删除文件
-     * @param url
-     * @param website
+     * 删除静态文件
+     * @param url 文件相对路径
+     * @param website 站点
      */
     public void removeFile(String url, Website website) {
         FileManager fileManager = fileManagerFactory.getFileManager(website.getDefaultFileManager().getId());
         Page page = this.pageService.findUniqueByPath(url,website.getId());
+        if(page==null){
+            return;
+        }
         List<PageItem> pageItems = page.getPageItems();
         for(PageItem pageItem : pageItems){
             fileManager.removeFile(pageItem.getFile());
         }
+    }
+
+    /**
+     * 预览
+     * @param url 页面相对路径
+     * @param websiteId 站点
+     */
+    public Map<String, Object> getPageDatas(String url, Long websiteId){
+        Page page = this.pageService.findUniqueByPath(url,websiteId);
+        return this.generate.getPageDataMap(page);
     }
 }
